bring prod up to date

+++ cloud-functions/utilities/purge/src/purge.go
@@ -11,12 +11,12 @@ import (
 	"regexp"
 	"strings"
 
-	"cloud.google.com/go/bigquery"
 	"cloud.google.com/go/datastore"
 )
 
 // ProjectID is the env var of project id
 var pid = os.Getenv("PROJECTID")
+var did = os.Getenv("DSPROJECTID")
 var env = os.Getenv("ENVIRONMENT")
 var uid = os.Getenv("CLIENTID")
 var pwd = os.Getenv("CLIENTSECRET")
@@ -24,12 +24,12 @@ var pwd = os.Getenv("CLIENTSECRET")
 // global vars
 var ctx context.Context
 var ds *datastore.Client
-var bq *bigquery.Client
+var fs *datastore.Client
 
 var allowedOperations = map[string]map[string]map[string]bool{
 	"datastore": map[string]map[string]bool{
-		"namespace": map[string]bool{"delete": true},
-		"kind":      map[string]bool{"delete": true},
+		"namespace": map[string]bool{"delete": true, "count": true},
+		"kind":      map[string]bool{"delete": true, "count": true},
 	},
 	"bigquery": map[string]map[string]bool{
 		"dataset": map[string]bool{"delete": true},
@@ -40,7 +40,7 @@ var allowedOperations = map[string]map[string]map[string]bool{
 func init() {
 	ctx = context.Background()
 	ds, _ = datastore.NewClient(ctx, pid)
-	bq, _ = bigquery.NewClient(ctx, pid)
+	fs, _ = datastore.NewClient(ctx, did)
 
 	log.Printf("init completed")
 }
@@ -97,9 +97,16 @@ func ProcessRequest(w http.ResponseWriter, r *http.Request) {
 				}
 
 				if strings.EqualFold(input.TargetType, "datastore") {
-					n, k, e := purgeDataStore(strings.ToLower(input.TargetLevel), input.TargetSelection, input.TargetSubSelection)
-					w.WriteHeader(http.StatusOK)
-					fmt.Fprintf(w, "{\"success\": true, \"message\": \"deleted %v namespaces, %v kinds, %v entities\"}", n, k, e)
+					if strings.ToLower(input.Operation) == "delete" {
+						n, k, e := purgeDataStore(strings.ToLower(input.TargetLevel), input.TargetSelection, input.TargetSubSelection)
+						w.WriteHeader(http.StatusOK)
+						fmt.Fprintf(w, "{\"success\": true, \"message\": \"deleted %v namespaces, %v kinds, %v entities\"}", n, k, e)
+					} else if strings.ToLower(input.Operation) == "count" {
+						n, k, e := countDataStore(strings.ToLower(input.TargetLevel), input.TargetSelection, input.TargetSubSelection)
+						w.WriteHeader(http.StatusOK)
+						fmt.Fprintf(w, "{\"success\": true, \"message\": \"counted %v namespaces, %v kinds, %v entities\"}", n, k, e)
+					}
+
 					return
 				}
 			} else {
@@ -120,14 +127,14 @@ func ProcessRequest(w http.ResponseWriter, r *http.Request) {
 
 }
 
-func purgeDataStore(level string, filter string, subfilter string) (int, int, int) {
+func countDataStore(level string, filter string, subfilter string) (int, int, int) {
 	countNS := 0
 	countKind := 0
 	countEntity := 0
 	switch level {
 	case "namespace":
 		query := datastore.NewQuery("__namespace__").KeysOnly()
-		namespaces, err := ds.GetAll(ctx, query, nil)
+		namespaces, err := fs.GetAll(ctx, query, nil)
 		if err != nil {
 			return countNS, countKind, countEntity
 		}
@@ -136,7 +143,73 @@ func purgeDataStore(level string, filter string, subfilter string) (int, int, in
 		for _, n := range namespaces {
 			if rens.MatchString(n.Name) {
 				query := datastore.NewQuery("__kind__").Namespace(n.Name).KeysOnly()
-				kinds, err := ds.GetAll(ctx, query, nil)
+				kinds, err := fs.GetAll(ctx, query, nil)
+				if err != nil {
+					return countNS, countKind, countEntity
+				}
+				if len(subfilter) > 0 {
+					rekind, _ := regexp.Compile(subfilter)
+					for _, k := range kinds {
+						if rekind.MatchString(k.Name) {
+							countKind++
+							countEntity += countDS(n.Name, k.Name)
+						}
+					}
+				} else {
+					for _, k := range kinds {
+						countKind++
+						countEntity += countDS(n.Name, k.Name)
+					}
+				}
+				countNS++
+			}
+		}
+	case "kind":
+		query := datastore.NewQuery("__kind__").Namespace(filter).KeysOnly()
+		if filter == "[default]" { // remove the namespace filter if [default]
+			query = datastore.NewQuery("__kind__").KeysOnly()
+		}
+		keys, err := fs.GetAll(ctx, query, nil)
+		if err != nil {
+			return countNS, countKind, countEntity
+		}
+		if len(subfilter) > 0 {
+
+			regex, _ := regexp.Compile(subfilter)
+			for _, k := range keys {
+				if regex.MatchString(k.Name) {
+					countKind++
+					countEntity += countDS(filter, k.Name)
+				}
+			}
+		} else {
+			for _, k := range keys {
+				countKind++
+				countEntity += countDS(filter, k.Name)
+			}
+		}
+		countNS++
+	}
+	return countNS, countKind, countEntity
+}
+
+func purgeDataStore(level string, filter string, subfilter string) (int, int, int) {
+	countNS := 0
+	countKind := 0
+	countEntity := 0
+	switch level {
+	case "namespace":
+		query := datastore.NewQuery("__namespace__").KeysOnly()
+		namespaces, err := fs.GetAll(ctx, query, nil)
+		if err != nil {
+			return countNS, countKind, countEntity
+		}
+		rens, _ := regexp.Compile("^" + env + "-" + filter)
+
+		for _, n := range namespaces {
+			if rens.MatchString(n.Name) {
+				query := datastore.NewQuery("__kind__").Namespace(n.Name).KeysOnly()
+				kinds, err := fs.GetAll(ctx, query, nil)
 				if err != nil {
 					return countNS, countKind, countEntity
 				}
@@ -162,7 +235,7 @@ func purgeDataStore(level string, filter string, subfilter string) (int, int, in
 		if filter == "[default]" { // remove the namespace filter if [default]
 			query = datastore.NewQuery("__kind__").KeysOnly()
 		}
-		keys, err := ds.GetAll(ctx, query, nil)
+		keys, err := fs.GetAll(ctx, query, nil)
 		if err != nil {
 			return countNS, countKind, countEntity
 		}
@@ -190,6 +263,19 @@ func purgeBigQuery(level string, filter string) {
 
 }
 
+func countDS(ns string, kind string) int {
+	if strings.HasPrefix(kind, "_") { // statistics entities, cannot delete them without error
+		return 0
+	}
+	query := datastore.NewQuery(kind).Namespace(ns).KeysOnly()
+	if ns == "[default]" {
+		query = datastore.NewQuery(kind).KeysOnly()
+	}
+	keys, _ := fs.GetAll(ctx, query, nil)
+
+	return len(keys)
+}
+
 func deleteDS(ns string, kind string) int {
 	if strings.HasPrefix(kind, "_") { // statistics entities, cannot delete them without error
 		return 0
@@ -198,7 +284,7 @@ func deleteDS(ns string, kind string) int {
 	if ns == "[default]" {
 		query = datastore.NewQuery(kind).KeysOnly()
 	}
-	keys, _ := ds.GetAll(ctx, query, nil)
+	keys, _ := fs.GetAll(ctx, query, nil)
 
 	l := len(keys) / 500
 	if len(keys)%500 > 0 {
@@ -214,7 +300,7 @@ func deleteDS(ns string, kind string) int {
 		log.Printf("Deleting records from ns %v, kind %v, %v - %v", ns, kind, s, e)
 		if r%10 == 0 {
 			log.Printf("Deleting records from ns %v, kind %v, %v - %v", ns, kind, s, e)
-			err := ds.DeleteMulti(ctx, keys[s:e])
+			err := fs.DeleteMulti(ctx, keys[s:e])
 
 			if err != nil {
 				log.Printf("Error Deleting records from ns %v, kind %v, err %v", ns, kind, err)
@@ -222,7 +308,7 @@ func deleteDS(ns string, kind string) int {
 		} else {
 			c := make(chan bool)
 			go func() {
-				err := ds.DeleteMulti(ctx, keys[s:e])
+				err := fs.DeleteMulti(ctx, keys[s:e])
 
 				if err != nil {
 					log.Printf("Error Deleting records from ns %v, kind %v, err %v", ns, kind, err)
